LayoutTests/fast/css-grid-layout/grid-item-column-row-get-set-expected.txt

Test that setting and getting grid-column and grid-row works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting -webkit-grid-column and -webkit-grid-row set through CSS
PASS getComputedStyle(gridItemWithNoCSSRule, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItemWithNoCSSRule, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItemWithNoCSSRule, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItemWithNoCSSRule, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItemWithNoCSSRule, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItemWithNoCSSRule, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItemWithPositiveInteger, '').getPropertyValue('-webkit-grid-column') is "10 / auto"
PASS getComputedStyle(gridItemWithPositiveInteger, '').getPropertyValue('-webkit-grid-column-start') is "10"
PASS getComputedStyle(gridItemWithPositiveInteger, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItemWithPositiveInteger, '').getPropertyValue('-webkit-grid-row') is "15 / auto"
PASS getComputedStyle(gridItemWithPositiveInteger, '').getPropertyValue('-webkit-grid-row-start') is "15"
PASS getComputedStyle(gridItemWithPositiveInteger, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItemWithNegativeInteger, '').getPropertyValue('-webkit-grid-column') is "-10 / auto"
PASS getComputedStyle(gridItemWithNegativeInteger, '').getPropertyValue('-webkit-grid-column-start') is "-10"
PASS getComputedStyle(gridItemWithNegativeInteger, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItemWithNegativeInteger, '').getPropertyValue('-webkit-grid-row') is "-15 / auto"
PASS getComputedStyle(gridItemWithNegativeInteger, '').getPropertyValue('-webkit-grid-row-start') is "-15"
PASS getComputedStyle(gridItemWithNegativeInteger, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItemWithAutoElement, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItemWithAutoElement, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItemWithAutoElement, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItemWithAutoElement, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItemWithAutoElement, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItemWithAutoElement, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItemWith2IntegerElement, '').getPropertyValue('-webkit-grid-column') is "10 / 15"
PASS getComputedStyle(gridItemWith2IntegerElement, '').getPropertyValue('-webkit-grid-column-start') is "10"
PASS getComputedStyle(gridItemWith2IntegerElement, '').getPropertyValue('-webkit-grid-column-end') is "15"
PASS getComputedStyle(gridItemWith2IntegerElement, '').getPropertyValue('-webkit-grid-row') is "5 / 5"
PASS getComputedStyle(gridItemWith2IntegerElement, '').getPropertyValue('-webkit-grid-row-start') is "5"
PASS getComputedStyle(gridItemWith2IntegerElement, '').getPropertyValue('-webkit-grid-row-end') is "5"
PASS getComputedStyle(gridItemWithNegativePositiveIntegerElement, '').getPropertyValue('-webkit-grid-column') is "10 / -10"
PASS getComputedStyle(gridItemWithNegativePositiveIntegerElement, '').getPropertyValue('-webkit-grid-column-start') is "10"
PASS getComputedStyle(gridItemWithNegativePositiveIntegerElement, '').getPropertyValue('-webkit-grid-column-end') is "-10"
PASS getComputedStyle(gridItemWithNegativePositiveIntegerElement, '').getPropertyValue('-webkit-grid-row') is "-8 / 5"
PASS getComputedStyle(gridItemWithNegativePositiveIntegerElement, '').getPropertyValue('-webkit-grid-row-start') is "-8"
PASS getComputedStyle(gridItemWithNegativePositiveIntegerElement, '').getPropertyValue('-webkit-grid-row-end') is "5"
PASS getComputedStyle(gridItemWithBeforeSpanElement, '').getPropertyValue('-webkit-grid-column') is "span 2 / 4"
PASS getComputedStyle(gridItemWithBeforeSpanElement, '').getPropertyValue('-webkit-grid-column-start') is "span 2"
PASS getComputedStyle(gridItemWithBeforeSpanElement, '').getPropertyValue('-webkit-grid-column-end') is "4"
PASS getComputedStyle(gridItemWithBeforeSpanElement, '').getPropertyValue('-webkit-grid-row') is "3 / span 5"
PASS getComputedStyle(gridItemWithBeforeSpanElement, '').getPropertyValue('-webkit-grid-row-start') is "3"
PASS getComputedStyle(gridItemWithBeforeSpanElement, '').getPropertyValue('-webkit-grid-row-end') is "span 5"
PASS getComputedStyle(gridItemWithAfterSpanElement, '').getPropertyValue('-webkit-grid-column') is "span 2 / 4"
PASS getComputedStyle(gridItemWithAfterSpanElement, '').getPropertyValue('-webkit-grid-column-start') is "span 2"
PASS getComputedStyle(gridItemWithAfterSpanElement, '').getPropertyValue('-webkit-grid-column-end') is "4"
PASS getComputedStyle(gridItemWithAfterSpanElement, '').getPropertyValue('-webkit-grid-row') is "3 / span 5"
PASS getComputedStyle(gridItemWithAfterSpanElement, '').getPropertyValue('-webkit-grid-row-start') is "3"
PASS getComputedStyle(gridItemWithAfterSpanElement, '').getPropertyValue('-webkit-grid-row-end') is "span 5"
PASS getComputedStyle(gridItemWith2OnlySpanElement, '').getPropertyValue('-webkit-grid-column') is "span 1 / span 1"
PASS getComputedStyle(gridItemWith2OnlySpanElement, '').getPropertyValue('-webkit-grid-column-start') is "span 1"
PASS getComputedStyle(gridItemWith2OnlySpanElement, '').getPropertyValue('-webkit-grid-column-end') is "span 1"
PASS getComputedStyle(gridItemWith2OnlySpanElement, '').getPropertyValue('-webkit-grid-row') is "span 1 / span 1"
PASS getComputedStyle(gridItemWith2OnlySpanElement, '').getPropertyValue('-webkit-grid-row-start') is "span 1"
PASS getComputedStyle(gridItemWith2OnlySpanElement, '').getPropertyValue('-webkit-grid-row-end') is "span 1"
PASS getComputedStyle(gridItemWith2AutoElement, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItemWith2AutoElement, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItemWith2AutoElement, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItemWith2AutoElement, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItemWith2AutoElement, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItemWith2AutoElement, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItemWithBothShortLongHandElement, '').getPropertyValue('-webkit-grid-column') is "10 / 11"
PASS getComputedStyle(gridItemWithBothShortLongHandElement, '').getPropertyValue('-webkit-grid-column-start') is "10"
PASS getComputedStyle(gridItemWithBothShortLongHandElement, '').getPropertyValue('-webkit-grid-column-end') is "11"
PASS getComputedStyle(gridItemWithBothShortLongHandElement, '').getPropertyValue('-webkit-grid-row') is "4 / 5"
PASS getComputedStyle(gridItemWithBothShortLongHandElement, '').getPropertyValue('-webkit-grid-row-start') is "4"
PASS getComputedStyle(gridItemWithBothShortLongHandElement, '').getPropertyValue('-webkit-grid-row-end') is "5"
PASS getComputedStyle(gridItemWithNoSpaceElement, '').getPropertyValue('-webkit-grid-column') is "auto / 1"
PASS getComputedStyle(gridItemWithNoSpaceElement, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItemWithNoSpaceElement, '').getPropertyValue('-webkit-grid-column-end') is "1"
PASS getComputedStyle(gridItemWithNoSpaceElement, '').getPropertyValue('-webkit-grid-row') is "5 / auto"
PASS getComputedStyle(gridItemWithNoSpaceElement, '').getPropertyValue('-webkit-grid-row-start') is "5"
PASS getComputedStyle(gridItemWithNoSpaceElement, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItemWithString, '').getPropertyValue('-webkit-grid-column') is "1 first / auto"
PASS getComputedStyle(gridItemWithString, '').getPropertyValue('-webkit-grid-column-start') is "1 first"
PASS getComputedStyle(gridItemWithString, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItemWithString, '').getPropertyValue('-webkit-grid-row') is "1 last / auto"
PASS getComputedStyle(gridItemWithString, '').getPropertyValue('-webkit-grid-row-start') is "1 last"
PASS getComputedStyle(gridItemWithString, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItemWithSpanString, '').getPropertyValue('-webkit-grid-column') is "1 / span 1 first"
PASS getComputedStyle(gridItemWithSpanString, '').getPropertyValue('-webkit-grid-column-start') is "1"
PASS getComputedStyle(gridItemWithSpanString, '').getPropertyValue('-webkit-grid-column-end') is "span 1 first"
PASS getComputedStyle(gridItemWithSpanString, '').getPropertyValue('-webkit-grid-row') is "-1 / span 1 last"
PASS getComputedStyle(gridItemWithSpanString, '').getPropertyValue('-webkit-grid-row-start') is "-1"
PASS getComputedStyle(gridItemWithSpanString, '').getPropertyValue('-webkit-grid-row-end') is "span 1 last"
PASS getComputedStyle(gridItemWithSpanNumberString, '').getPropertyValue('-webkit-grid-column') is "1 / span 3 first"
PASS getComputedStyle(gridItemWithSpanNumberString, '').getPropertyValue('-webkit-grid-column-start') is "1"
PASS getComputedStyle(gridItemWithSpanNumberString, '').getPropertyValue('-webkit-grid-column-end') is "span 3 first"
PASS getComputedStyle(gridItemWithSpanNumberString, '').getPropertyValue('-webkit-grid-row') is "-1 / span 2 last"
PASS getComputedStyle(gridItemWithSpanNumberString, '').getPropertyValue('-webkit-grid-row-start') is "-1"
PASS getComputedStyle(gridItemWithSpanNumberString, '').getPropertyValue('-webkit-grid-row-end') is "span 2 last"
PASS getComputedStyle(gridItemWithSingleNamedGridArea, '').getPropertyValue('-webkit-grid-column') is "thirdArea / thirdArea"
PASS getComputedStyle(gridItemWithSingleNamedGridArea, '').getPropertyValue('-webkit-grid-column-start') is "thirdArea"
PASS getComputedStyle(gridItemWithSingleNamedGridArea, '').getPropertyValue('-webkit-grid-column-end') is "thirdArea"
PASS getComputedStyle(gridItemWithSingleNamedGridArea, '').getPropertyValue('-webkit-grid-row') is "firstArea / firstArea"
PASS getComputedStyle(gridItemWithSingleNamedGridArea, '').getPropertyValue('-webkit-grid-row-start') is "firstArea"
PASS getComputedStyle(gridItemWithSingleNamedGridArea, '').getPropertyValue('-webkit-grid-row-end') is "firstArea"
PASS getComputedStyle(gridItemWithNamedGridAreaAndSpan, '').getPropertyValue('-webkit-grid-column') is "thirdArea / span 1"
PASS getComputedStyle(gridItemWithNamedGridAreaAndSpan, '').getPropertyValue('-webkit-grid-column-start') is "thirdArea"
PASS getComputedStyle(gridItemWithNamedGridAreaAndSpan, '').getPropertyValue('-webkit-grid-column-end') is "span 1"
PASS getComputedStyle(gridItemWithNamedGridAreaAndSpan, '').getPropertyValue('-webkit-grid-row') is "firstArea / span 2"
PASS getComputedStyle(gridItemWithNamedGridAreaAndSpan, '').getPropertyValue('-webkit-grid-row-start') is "firstArea"
PASS getComputedStyle(gridItemWithNamedGridAreaAndSpan, '').getPropertyValue('-webkit-grid-row-end') is "span 2"

Test the initial value
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-column') is 'auto / auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-column-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-column-end') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-row') is 'auto / auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-row-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-row-end') is 'auto'

Test getting and setting grid-column and grid-row through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "18 / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "18"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "66 / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "66"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "-55 / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "-55"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "-40 / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "-40"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "10 / 55"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "10"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "55"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "1 / 10"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "1"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "10"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "span 5 / 5"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "span 5"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "5"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "4 / span 4"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "4"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "span 4"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "-5 / 5"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "-5"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "5"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "4 / -4"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "4"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "-4"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "4 / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "4"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "5 / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "5"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / 5"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "5"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / 8"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "8"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "span 1 / 3"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "span 1"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "3"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "5 / span 1"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "5"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "span 1"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "span 1 first / 3"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "span 1 first"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "3"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "5 / span 1 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "5"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "span 1 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "1 first / 1 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "1 first"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "1 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "1 nav / span 1 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "1 nav"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "span 1 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "3 first / 2 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "3 first"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "2 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "5 nav / span 7 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "5 nav"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "span 7 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "span 3 first / -3 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "span 3 first"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "-3 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "span 2 last / -1 nav"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "span 2 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "-1 nav"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "5 / none"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "5"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "none"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "8 / foobar"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "8"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "foobar"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "nonExistent / none"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "nonExistent"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "none"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "nonExistent / foobar"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "nonExistent"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "foobar"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "span 3 first / none"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "span 3 first"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "none"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "span 1 last / foobar"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "span 1 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "foobar"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "span 5 / span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "span 5"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "span 1 first / span 1 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "span 1 first"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "span 1 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "span 5 first / span 2 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "span 5 first"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "span 2 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "span 3 first / span 7 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "span 3 first"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "span 7 last"

Test setting grid-column and grid-row back to 'auto' through JS
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-column') is '18 / 19'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-column-start') is '18'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-column-end') is '19'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-row') is '66 / 68'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-row-start') is '66'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-row-end') is '68'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-column') is 'auto / auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-column-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-column-end') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-row') is 'auto / auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-row-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-row-end') is 'auto'

Test getting and setting 'initial' grid-column and grid-row through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "1 / span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "1"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "span 3 / 5"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "span 3"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "5"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "span 3 / 5"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "span 3"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "5"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"

Test getting and setting 'inherit' grid-column and grid-row through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "1 / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "1"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "span 1 / 7"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "span 1"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "7"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "6 / span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "6"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "1 / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "1"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "6 / span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "6"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "span 1 / 7"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "span 1"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "7"

Test getting and setting invalid grid-column and grid-row through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row-end') is "auto"
PASS successfullyParsed is true

TEST COMPLETE
